MBS-10577 Rename build-checks plugin package
