+++ subprojects/gradle/build-checks/src/main/kotlin/com/avito/android/build_checks/internal/CheckResult.kt
@@ -1,0 +1,6 @@
+package com.avito.android.build_checks.internal
+
+internal sealed class CheckResult {
+    object Ok : CheckResult()
+    class Failed(val message: String) : CheckResult()
+}
