+++ subprojects/gradle/build-checks/src/main/kotlin/com/avito/android/build_checks/internal/FailedCheckMessage.kt
@@ -1,0 +1,28 @@
+package com.avito.android.build_checks.internal
+
+import com.avito.android.build_checks.extensionName
+import kotlin.reflect.KFunction
+
+internal class FailedCheckMessage(
+    /**
+     * Reference to the [BuildChecksExtension] method
+     */
+    extensionMethodReference: KFunction<Any>,
+    private val message: String
+) {
+
+    private val checkExtensionName = extensionMethodReference.name
+
+    override fun toString() =
+        """
+ERROR: '$checkExtensionName' build check is failed.
+${message.trimIndent()}
+This check can be disabled by extension:
+$extensionName {
+    $checkExtensionName {
+        enabled = false
+    }
+}
+See https://avito-tech.github.io/avito-android/docs/projects/buildchecks/ (search '$checkExtensionName')
+""".trimIndent()
+}
