+++ subprojects/gradle/build-checks/src/main/kotlin/com/avito/android/build_checks/internal/incremental_kapt/PrintStreamMessager.kt
@@ -1,0 +1,37 @@
+package com.avito.android.build_checks.internal.incremental_kapt
+
+import java.io.PrintStream
+import javax.annotation.processing.Messager
+import javax.lang.model.element.AnnotationMirror
+import javax.lang.model.element.AnnotationValue
+import javax.lang.model.element.Element
+import javax.tools.Diagnostic
+
+internal class PrintStreamMessager(
+    private val writer: PrintStream = System.out,
+    private val tag: String = ""
+) : Messager {
+
+    override fun printMessage(kind: Diagnostic.Kind?, msg: CharSequence?) {
+        val prefix = if (tag.isEmpty()) "" else "$tag: "
+        writer.println("$prefix($kind) $msg")
+    }
+
+    override fun printMessage(kind: Diagnostic.Kind?, msg: CharSequence?, e: Element?) {
+        printMessage(kind, msg)
+    }
+
+    override fun printMessage(kind: Diagnostic.Kind?, msg: CharSequence?, e: Element?, am: AnnotationMirror?) {
+        printMessage(kind, msg)
+    }
+
+    override fun printMessage(
+        kind: Diagnostic.Kind?,
+        msg: CharSequence?,
+        element: Element?,
+        mirror: AnnotationMirror?,
+        value: AnnotationValue?
+    ) {
+        printMessage(kind, msg)
+    }
+}
