+++ subprojects/gradle/build-checks/src/main/kotlin/com/avito/android/build_checks/internal/incremental_kapt/StubProcessingEnvironment.kt
@@ -1,0 +1,31 @@
+package com.avito.android.build_checks.internal.incremental_kapt
+
+import java.util.Locale
+import javax.annotation.processing.Filer
+import javax.annotation.processing.Messager
+import javax.annotation.processing.ProcessingEnvironment
+import javax.lang.model.SourceVersion
+import javax.lang.model.util.Elements
+import javax.lang.model.util.Types
+
+internal class StubProcessingEnvironment(
+    private val options: Map<String, String> = emptyMap(),
+    private val logTag: String = ""
+) : ProcessingEnvironment {
+
+    override fun getElementUtils(): Elements? = supposeNotUsed()
+
+    override fun getTypeUtils(): Types? = supposeNotUsed()
+
+    override fun getMessager(): Messager = PrintStreamMessager(tag = logTag)
+
+    override fun getLocale(): Locale = Locale.getDefault()
+
+    override fun getSourceVersion(): SourceVersion = SourceVersion.RELEASE_8
+
+    override fun getOptions(): MutableMap<String, String> = options.toMutableMap()
+
+    override fun getFiler(): Filer? = supposeNotUsed()
+
+    private fun <T> supposeNotUsed(): T? = null
+}
