+++ subprojects/gradle/build-checks/src/main/kotlin/com/avito/android/build_checks/internal/params/ParamMismatchFailure.kt
@@ -1,0 +1,3 @@
+package com.avito.android.build_checks.internal.params
+
+internal class ParamMismatchFailure(cause: Throwable) : Exception(cause)
