--- subprojects/gradle/build-checks/src/main/kotlin/com/avito/android/plugin/build_param_check/CheckResult.kt
@@ -1,6 +1,0 @@
-package com.avito.android.plugin.build_param_check
-
-internal sealed class CheckResult {
-    object Ok : CheckResult()
-    class Failed(val message: String) : CheckResult()
-}
