--- subprojects/gradle/build-checks/src/main/kotlin/com/avito/android/plugin/build_param_check/ParamMismatchFailure.kt
@@ -1,3 +1,0 @@
-package com.avito.android.plugin.build_param_check
-
-internal class ParamMismatchFailure(cause: Throwable) : Exception(cause)
